Made some updates for getPersons test, chanced id from int to long

+++ ca2-backend/src/main/java/com/mycompany/ca2/backend/facade/implementations/FacadeImp.java
@@ -84,7 +84,7 @@ public class FacadeImp implements Facade{
     }
 
     @Override
-    public Person addHobbyToPerson(int personId, Hobby hobby) {
+    public Person addHobbyToPerson(Long personId, Hobby hobby) {
         return this.infoEntityFacade.addHobbyToPerson(personId, hobby);
     }
