+++ ca2-backend/src/main/java/com/mycompany/ca2/backend/facade/interfaces/Facade.java
@@ -94,7 +94,7 @@ public interface Facade {
      * @param personId
      * @return 
      */
-    Person addHobbyToPerson(int personId, Hobby hobby);
+    Person addHobbyToPerson(Long personId, Hobby hobby);
     
     /**
      * Adds a new Phone object to the InfoEntity with the given Id
